refactor: split secure domain profile resolution
Constraint: Preserve stable manifest checks, fail-closed resolution, digest bytes, and the 64-match ceiling.
Rejected: Replace descriptor-bound reads with path reads | would weaken store-swap and symlink defenses.
Confidence: high
Scope-risk: narrow
Directive: Keep profile snapshot validation descriptor-bound and preserve the query compatibility seam.
Tested: 256 domain, routing, project-context, wrapper golden, and wrapper-contract tests; targeted 32-test rerun; compileall; Ruff check and format; git diff --check.
Not-tested: Full repository unittest discovery.
