+++ src/workflows/domain_routing_context.py
@@ -1,10 +1,25 @@
 from __future__ import annotations
 
-from dataclasses import dataclass
 import hashlib
 import json
 from typing import Sequence
 import unicodedata
+
+from .domain_intelligence_profile_resolution import (
+    MAX_DOMAIN_CONTEXT_MATCHES,
+    DomainClarificationTarget,
+    matches_reviewed_phrase,
+    resolve_domain_clarification_target,
+)
+
+
+__all__ = (
+    "MAX_DOMAIN_CONTEXT_MATCHES",
+    "DomainClarificationTarget",
+    "build_domain_routing_context",
+    "matches_reviewed_phrase",
+    "resolve_domain_routing_context",
+)
 
 
 DOMAIN_ROUTING_CONTEXT_KEY = "domain_routing_context"
@@ -19,16 +34,7 @@ MAX_WORKFLOW_HINT_CODE_POINTS = 120
 MAX_REQUIRED_INPUT_CODE_POINTS = 120
 MAX_QUESTION_CODE_POINTS = 240
 MAX_CLAIM_BOUNDARY_CODE_POINTS = 320
-MAX_DOMAIN_CONTEXT_MATCHES = 64
 _SUPPORTED_QUESTION_LOCALES = frozenset({"en", "ko"})
-
-
-@dataclass(frozen=True)
-class DomainClarificationTarget:
-    workflow_hint: str
-    required_input: str
-    question_locale: str
-    question_text: str
 
 
 def build_domain_routing_context(
@@ -55,32 +61,6 @@ def build_domain_routing_context(
     return {DOMAIN_ROUTING_CONTEXT_KEY: context}
 
 
-def matches_reviewed_phrase(message: object, phrase: object) -> bool:
-    """Return whether a normalized literal phrase has a valid Unicode boundary match."""
-    normalized_message = _normalize_match_text(message)
-    normalized_phrase = _normalize_match_text(phrase)
-    if not normalized_message or not normalized_phrase:
-        return False
-
-    offset = normalized_message.find(normalized_phrase)
-    while offset >= 0:
-        end = offset + len(normalized_phrase)
-        left_valid = (
-            not _is_word_like(normalized_phrase[0])
-            or offset == 0
-            or not _is_word_like(normalized_message[offset - 1])
-        )
-        right_valid = (
-            not _is_word_like(normalized_phrase[-1])
-            or end == len(normalized_message)
-            or not _is_word_like(normalized_message[end])
-        )
-        if left_valid and right_valid:
-            return True
-        offset = normalized_message.find(normalized_phrase, offset + 1)
-    return False
-
-
 def resolve_domain_routing_context(
     binding: object,
     message: object,
@@ -88,8 +68,6 @@ def resolve_domain_routing_context(
     locale: str,
 ) -> dict[str, object] | None:
     """Resolve one catalog-owned question from a complete reviewed store snapshot."""
-    from ..paths import project_identity
-    from ..skills import catalog
     from .domain_intelligence_queries import read_validated_domain_profiles_at
     from .domain_project_context import HostProjectBinding
 
@@ -97,63 +75,14 @@ def resolve_domain_routing_context(
         return None
     try:
         profiles = read_validated_domain_profiles_at(binding)
-        expected_scope = {
-            "kind": "project",
-            "ref": project_identity(binding.project_root),
-            "ref_authority": "operator_or_wrapper_supplied",
-            "identity_claim": "not_authenticated_identity_evidence",
-        }
-        matches: list[tuple[dict[str, object], dict[str, object]]] = []
-        for profile in profiles:
-            if profile.get("scope") != expected_scope:
-                continue
-            mappings = profile.get("vocabulary_mappings")
-            if not isinstance(mappings, list):
-                return None
-            for mapping in mappings:
-                if not isinstance(mapping, dict):
-                    return None
-                if matches_reviewed_phrase(message, mapping.get("phrase")):
-                    matches.append((profile, mapping))
-                    if len(matches) > MAX_DOMAIN_CONTEXT_MATCHES:
-                        return None
-        if not matches:
-            return None
-
-        canonicals = {str(mapping.get("canonical")) for _profile, mapping in matches}
-        if len(canonicals) != 1:
-            return None
-
-        routable = {
-            definition.name: definition for definition in catalog.routable_definitions()
-        }
-        selected_hints: set[str] = set()
-        for profile, _mapping in matches:
-            hints = profile.get("workflow_hints")
-            if not isinstance(hints, list) or not hints:
-                return None
-            for hint in hints:
-                if not isinstance(hint, str) or hint not in routable:
-                    return None
-                selected_hints.add(hint)
-        if len(selected_hints) != 1:
-            return None
-
-        workflow_hint = next(iter(selected_hints))
-        definition = routable[workflow_hint]
-        questions = definition.expert_questions
-        if not questions:
-            return None
-        question = questions[0]
-        if question.required_input not in definition.required_inputs:
-            return None
-        selected_locale = "ko" if locale == "ko" else "en"
-        target = DomainClarificationTarget(
-            workflow_hint=workflow_hint,
-            required_input=question.required_input,
-            question_locale=selected_locale,
-            question_text=question.ko if selected_locale == "ko" else question.en,
+        target = resolve_domain_clarification_target(
+            profiles,
+            message,
+            project_root=binding.project_root,
+            locale=locale,
         )
+        if target is None:
+            return None
         return build_domain_routing_context((target,))
     except (OSError, TypeError, ValueError):
         return None
@@ -187,14 +116,3 @@ def _canonical_public_digest(context: dict[str, object]) -> str:
         ensure_ascii=False,
     )
     return hashlib.sha256(preimage.encode("utf-8")).hexdigest()
-
-
-def _normalize_match_text(value: object) -> str:
-    if not isinstance(value, str):
-        return ""
-    normalized = unicodedata.normalize("NFKC", value).strip()
-    return " ".join(normalized.split()).casefold()
-
-
-def _is_word_like(value: str) -> bool:
-    return value == "_" or unicodedata.category(value)[0] in {"L", "N", "M"}
